解决 .podspec 文件校验的警告

+++ TWTool/Classes/TWDatePicker/TWDatePicker.m
@@ -540,11 +540,11 @@
             if (component == 0) {
                 yearIndex = row;
                 self.showYearLabel.text =_yearArray[yearIndex];
-                NSLog(@"yearIndex = %ld",row);
+                NSLog(@"yearIndex = %ld", (long)row);
             }
             if (component == 1) {
                 monthIndex = row;
-                NSLog(@"monthIndex = %ld",row);
+                NSLog(@"monthIndex = %ld",(long)row);
             }
         }
             break;
